+++ mago3d-user/src/main/java/gaia3d/service/impl/LayerServiceImpl.java
@@ -1,13 +1,12 @@
 package gaia3d.service.impl;
 
-import java.util.List;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
 import gaia3d.domain.layer.Layer;
 import gaia3d.persistence.LayerMapper;
 import gaia3d.service.LayerService;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
 
 /**
  * 여기서는 Geoserver Rest API 결과를 가지고 파싱 하기 때문에 RestTemplate을 커스트마이징하면 안됨
@@ -37,8 +36,8 @@ public class LayerServiceImpl implements LayerService {
      * @return
      */
     @Transactional(readOnly=true)
-    public Layer getLayer(Long layerId) {
-        return layerMapper.getLayer(layerId.intValue());
+    public Layer getLayer(Integer layerId) {
+        return layerMapper.getLayer(layerId);
     }
 
     /**
